clean up session launcher buttons
